[#105] Moved Hub_Multiple to a dedicated unit test file, now including presets
